New algorithm
Rewriting the graph search code based on the new algorithm.
Added graphics functions and abilities.
Issues:
Connecting edges with nodes on the s-edges of the s-theta planes is problematic, the current patch is to ignore the issue, and not connect beyond the edge (shutting down the s-cycle).
This code is relatively untested, and unproven on more complex objects.